/ can now make searches on /database
