+++ API/controller/routes/database.js
@@ -14,6 +14,7 @@ router.use('/meteorite', singleBodyRouter);
 
 /* GET database page. */
 router.get('/', async (req, res, next) => {
+  //console.log(req);
   let resObj = [];
   try {
     const Entries = db.aQuery('SELECT * FROM complete_table', []);
@@ -21,7 +22,7 @@ router.get('/', async (req, res, next) => {
   } catch (err) {
     next(createError(500));
   } finally {
-    console.log(resObj[0].rows);
+    // console.log(resObj[0].rows);
     res.send({
       Entries: resObj[0].rows,
     });
