+++ client/src/components/pages/Main.jsx
@@ -57,7 +57,8 @@ function Main(props) {
             <Route path='/register' component={Register}/>
             <ProtectedRoute path="/account" component={Account}/>
 
-            <ProtectedRoute path="/database" component={Database}/>
+            <Route path="/database" component={Database}/>
+            
             
             <ProtectedRoute path="/panel" component={Panel}/>
             <ProtectedRoute path="/data-entry" component={DataEntry}/>
