+++ client/src/components/styles/Home.scss
@@ -11,8 +11,6 @@ h3 {
 	padding-right: auto;
 }
 
-
-
 p {
 	font-size: 0.4rem;
 }
@@ -22,6 +20,8 @@ p {
 	-moz-box-shadow: 0px 4px 12px 0px rgba(0,0,0,0.78);
 	box-shadow: 0px 4px 12px 0px rgba(0,0,0,0.78);
 	background-color: #592651;
+	height: 72px;
+	padding: 17px;
 }
 
 #top-container {
